split pool details subcomponents
